requirements.txt
textual
sqlalchemy
google-generativeai